refactor: ProjectDash  posts to own component: ProjectDashPosts
